requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.10
django-cors-headers==3.4.0
django-phonenumber-field==5.0.0
django-tinymce==3.1.0
django-twilio==0.13.0.post1
djangorestframework==3.11.1
idna==2.10
phonenumbers==8.12.9
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
twilio==6.45.1
urllib3==1.25.10
